+++ src/models/Event/Activity/Biking.ts
@@ -1,0 +1,14 @@
+import { Schema } from "mongoose";
+import { activitySchema } from ".";
+import { Activity, Biking } from "@/types/Event";
+import { hikingSchema } from "./Hiking";
+import { EventModel } from "..";
+
+const bikingSchema = new Schema<Activity<Biking>>({
+  ...activitySchema.obj,
+  ...hikingSchema.obj,
+});
+
+export const BikingModel =
+  EventModel.discriminators?.Biking ||
+  EventModel.discriminator<Activity<Biking>>("Biking", bikingSchema);
